alx_travel_app/requirement.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
drf-yasg==1.21.8
python-decouple==3.8